requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
argh==0.26.2              # via watchdog
chardet==3.0.4            # via requests
click==6.7
gitdb2==2.0.3             # via gitpython
gitpython==2.1.9
idna==2.6                 # via requests
pathtools==0.1.2          # via watchdog
pyyaml==3.12
requests==2.18.4
smmap2==2.0.3             # via gitdb2
urllib3==1.22             # via requests
watchdog==0.8.3
